pyproject.toml
[project]
name = "langchain-mcp-deepeval-trajectory-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.5",
    "aiohttp-sse-client>=0.2.1",
    "fastmcp>=3.2.3",
    "langchain>=1.2.15",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langgraph-cli[inmem]>=0.4.21",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent_call", "client_apis"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_call*", "client_apis*"]
exclude = ["demo*", "langchain_learn*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top (dotenv loading order)
    "B008",   # do not perform function calls in argument defaults
    "SIM102", # collapsible if statements
    "SIM108", # use ternary
    "RUF001", # ambiguous character in string (Chinese text intentional)
    "RUF002", # ambiguous character in docstring (Chinese text intentional)
    "RUF003", # ambiguous character in comment (Chinese text intentional)
]

[tool.ruff.lint.isort]
known-first-party = ["agent_call", "client_apis"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
